RichPresence: update RPC after asset list retrieved
